+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # LAST CHANGES:
 
 
-Version `1.1.0a3 (5 Jan'22)`
+Version `1.1.0a4 (5 Jan'22)`
 
   - BUGFIX: Tab key (command completion) in kubuntu, reported and fixed by Steve
   - BUGFIX: saving modified files by "Preferences/Save", reported by Steve
   - NEW   : de.msg by Holger
   - NEW   : acknoledgements tab in About
   - CHANGE: port to listen made customizable, as suggested by Holger
+  - CHANGE: no scrollbar in "Complete Command" window
   - CHANGE: apave package